@@ -10,7 +10,6 @@ using Apps.Airtable.UrlBuilders;
 using Blackbird.Applications.Sdk.Common.Actions;
 using Blackbird.Applications.Sdk.Common.Invocation;
 using Newtonsoft.Json;
-using File = Blackbird.Applications.Sdk.Common.Files.File;
 
 namespace Apps.Airtable.Actions;
 
@@ -41,54 +40,55 @@ public class RecordActions : BaseInvocable
 
     [Action("Get value of string field", Description = "Get the value of a string field (e.g. single line text, " +
                                                        "long text, phone number, email, URL, single select).")]
-    public async Task<FieldValueResponse<string>> GetStringFieldValue([ActionParameter] TableIdentifier tableIdentifier, 
+    public async Task<FieldValueResponse<string>> GetStringFieldValue([ActionParameter] TableIdentifier tableIdentifier,
         [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier)
     {
         var field = await GetFieldValue(tableIdentifier, recordIdentifier, fieldIdentifier);
         return new FieldValueResponse<string> { Value = field?.ToString() ?? string.Empty };
     }
-    
-    [Action("Get value of number field", Description = "Get the value of a number field (e.g. number, currency, percent, " +
-                                                       "rating).")]
-    public async Task<FieldValueResponse<double?>> GetNumberFieldValue([ActionParameter] TableIdentifier tableIdentifier, 
+
+    [Action("Get value of number field", Description =
+        "Get the value of a number field (e.g. number, currency, percent, " +
+        "rating).")]
+    public async Task<FieldValueResponse<double?>> GetNumberFieldValue(
+        [ActionParameter] TableIdentifier tableIdentifier,
         [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier)
     {
         var field = await GetFieldValue(tableIdentifier, recordIdentifier, fieldIdentifier);
         return new FieldValueResponse<double?> { Value = (double?)field };
     }
-    
+
     [Action("Get value of date field", Description = "Get the value of a date field.")]
-    public async Task<FieldValueResponse<DateTimeOffset?>> GetDateFieldValue([ActionParameter] TableIdentifier tableIdentifier, 
+    public async Task<FieldValueResponse<DateTimeOffset?>> GetDateFieldValue(
+        [ActionParameter] TableIdentifier tableIdentifier,
         [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier)
     {
         var field = await GetFieldValue(tableIdentifier, recordIdentifier, fieldIdentifier);
         return new FieldValueResponse<DateTimeOffset?> { Value = (DateTime?)field };
     }
-    
+
     [Action("Get value of boolean field", Description = "Get the value of a boolean field (e.g. checkbox).")]
-    public async Task<FieldValueResponse<bool>> GetBooleanFieldValue([ActionParameter] TableIdentifier tableIdentifier, 
+    public async Task<FieldValueResponse<bool>> GetBooleanFieldValue([ActionParameter] TableIdentifier tableIdentifier,
         [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier)
     {
         var field = await GetFieldValue(tableIdentifier, recordIdentifier, fieldIdentifier);
         return new FieldValueResponse<bool> { Value = field != null };
     }
-    
+
     [Action("Download files from attachment field", Description = "Download files from an attachment field.")]
-    public async Task<FilesResponse> DownloadFilesFromAttachmentField([ActionParameter] TableIdentifier tableIdentifier, 
+    public async Task<FilesResponse> DownloadFilesFromAttachmentField([ActionParameter] TableIdentifier tableIdentifier,
         [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier)
     {
         var field = await GetFieldValue(tableIdentifier, recordIdentifier, fieldIdentifier);
         var files = JsonConvert.DeserializeObject<IEnumerable<FileDto>>(field?.ToString() ?? string.Empty,
             _jsonSerializerSettings) ?? new FileDto[] { };
         var downloadedFiles = new List<FileWrapper>();
-        
+
         foreach (var file in files)
         {
-            var client = new RestClient(file.Url);
-            var response = await client.ExecuteAsync(new RestRequest(""));
             downloadedFiles.Add(new FileWrapper
             {
-                File = new File(response.RawBytes)
+                File = new(new(HttpMethod.Get, file.Url))
                 {
                     Name = file.Filename,
                     ContentType = file.Type
@@ -99,11 +99,11 @@ public class RecordActions : BaseInvocable
         return new FilesResponse { Files = downloadedFiles };
     }
 
-    private async Task<object?> GetFieldValue(TableIdentifier tableIdentifier, RecordIdentifier recordIdentifier, 
+    private async Task<object?> GetFieldValue(TableIdentifier tableIdentifier, RecordIdentifier recordIdentifier,
         FieldIdentifier fieldIdentifier)
     {
         await CheckIfFieldExistsInTable(tableIdentifier, fieldIdentifier);
-        var request = new AirtableRequest($"/{tableIdentifier.TableId}/{recordIdentifier.RecordId}", Method.Get, 
+        var request = new AirtableRequest($"/{tableIdentifier.TableId}/{recordIdentifier.RecordId}", Method.Get,
             _credentials);
 
         try
@@ -117,16 +117,18 @@ public class RecordActions : BaseInvocable
             throw new Exception("Record with Record ID provided was not found.");
         }
     }
-    
+
     #endregion
 
     #region PATCH
 
-    [Action("Update value of string field", Description = "Update the value of a string field (e.g. single line text, " +
-                                                          "long text, phone number, email, URL, single select).")]
-    public async Task<RecordDto> UpdateStringFieldValue([ActionParameter] TableIdentifier tableIdentifier, 
-        [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier, 
-        [ActionParameter] [Display("New value")] string newValue)
+    [Action("Update value of string field", Description =
+        "Update the value of a string field (e.g. single line text, " +
+        "long text, phone number, email, URL, single select).")]
+    public async Task<RecordDto> UpdateStringFieldValue([ActionParameter] TableIdentifier tableIdentifier,
+        [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier,
+        [ActionParameter] [Display("New value")]
+        string newValue)
     {
         var jsonBody = $@"
         {{
@@ -137,12 +139,14 @@ public class RecordActions : BaseInvocable
         var record = await UpdateFieldValue(tableIdentifier, recordIdentifier, fieldIdentifier, jsonBody);
         return record;
     }
-    
-    [Action("Update value of number field", Description = "Update the value of a number field (e.g. number, currency, " +
-                                                          "percent, rating).")]
-    public async Task<RecordDto> UpdateNumberFieldValue([ActionParameter] TableIdentifier tableIdentifier, 
-        [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier, 
-        [ActionParameter] [Display("New value")] double newValue)
+
+    [Action("Update value of number field", Description =
+        "Update the value of a number field (e.g. number, currency, " +
+        "percent, rating).")]
+    public async Task<RecordDto> UpdateNumberFieldValue([ActionParameter] TableIdentifier tableIdentifier,
+        [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier,
+        [ActionParameter] [Display("New value")]
+        double newValue)
     {
         var jsonBody = $@"
         {{
@@ -155,9 +159,10 @@ public class RecordActions : BaseInvocable
     }
 
     [Action("Update value of date field", Description = "Update the value of a date field.")]
-    public async Task<RecordDto> UpdateDateFieldValue([ActionParameter] TableIdentifier tableIdentifier, 
-        [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier, 
-        [ActionParameter] [Display("New value")] DateTime newValue)
+    public async Task<RecordDto> UpdateDateFieldValue([ActionParameter] TableIdentifier tableIdentifier,
+        [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier,
+        [ActionParameter] [Display("New value")]
+        DateTime newValue)
     {
         var jsonBody = $@"
         {{
@@ -172,7 +177,8 @@ public class RecordActions : BaseInvocable
     [Action("Update value of boolean field", Description = "Update the value of a boolean field (e.g. checkbox).")]
     public async Task<RecordDto> UpdateBooleanFieldValue([ActionParameter] TableIdentifier tableIdentifier,
         [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier,
-        [ActionParameter] [Display("New value")] bool newValue)
+        [ActionParameter] [Display("New value")]
+        bool newValue)
     {
         var jsonBody = $@"
         {{
@@ -183,9 +189,9 @@ public class RecordActions : BaseInvocable
         var record = await UpdateFieldValue(tableIdentifier, recordIdentifier, fieldIdentifier, jsonBody);
         return record;
     }
-    
+
     //[Action("Upload file to attachment field", Description = "Upload a file to an attachment field.")]
-    public async Task<RecordDto> UploadFileToAttachmentField([ActionParameter] TableIdentifier tableIdentifier, 
+    public async Task<RecordDto> UploadFileToAttachmentField([ActionParameter] TableIdentifier tableIdentifier,
         [ActionParameter] RecordIdentifier recordIdentifier, [ActionParameter] FieldIdentifier fieldIdentifier,
         [ActionParameter] FileRequest file)
     {
@@ -204,22 +210,22 @@ public class RecordActions : BaseInvocable
         }
 
         jsonBody.AppendLine("{");
-        jsonBody.AppendLine($"\"url\": \"{file.File.DownloadUrl}\",");
+        jsonBody.AppendLine($"\"url\": \"{file.File.Url}\",");
         jsonBody.AppendLine($"\"filename\": \"{file.File.Name}\"");
         jsonBody.AppendLine("}");
         jsonBody.AppendLine("]");
         jsonBody.AppendLine("}");
         jsonBody.AppendLine("}");
-        
+
         var record = await UpdateFieldValue(tableIdentifier, recordIdentifier, fieldIdentifier, jsonBody.ToString());
         return record;
     }
 
-    private async Task<RecordDto> UpdateFieldValue(TableIdentifier tableIdentifier, RecordIdentifier recordIdentifier, 
+    private async Task<RecordDto> UpdateFieldValue(TableIdentifier tableIdentifier, RecordIdentifier recordIdentifier,
         FieldIdentifier fieldIdentifier, string jsonBody)
     {
         await CheckIfFieldExistsInTable(tableIdentifier, fieldIdentifier);
-        var request = new AirtableRequest($"/{tableIdentifier.TableId}/{recordIdentifier.RecordId}", Method.Patch, 
+        var request = new AirtableRequest($"/{tableIdentifier.TableId}/{recordIdentifier.RecordId}", Method.Patch,
             _credentials);
         request.AddJsonBody(jsonBody);
 
@@ -235,7 +241,7 @@ public class RecordActions : BaseInvocable
     }
 
     #endregion
-    
+
     private async Task CheckIfFieldExistsInTable(TableIdentifier tableIdentifier, FieldIdentifier fieldIdentifier)
     {
         var client = new AirtableClient(_credentials, new AirtableMetaUrlBuilder());
